Removes duplicate run of freshclam.
If the freshclam service is enabled, freshclam will have already begun
it's initial definition download. Forcing an execution of freshclam at
this point can collide with an already running instance.

Let's move this forced execution to the case where the user doesn't want
freshclam service enabled, but we need to ensure the virus definitions
are updated at least once.